feature : scrap 도메인 user 권한 설정 (#41)
* feature : 검증 로직 구현 및 분리

* feature : 임시 코드 제거 및 @AuthenticationPrincipal 적용

* feature : scrap 조회 기능 @AuthenticationPrincipal 적용(비공개 전환)

* refactor : scrap service refactoring

+++ src/main/java/com/practice/OAuth2/domain/scrap/service/ScrapService.java
@@ -24,7 +24,7 @@ public class ScrapService {
     private final ScrapRepository scrapRepository;
     private final UserRepository userRepository;
 
-    // scrap folder 생성
+    // 스크랩 폴더 생성
     @Transactional
     public Long createScrapFolder(Long userId, ScrapReqest.CreateScrapFolder req){
         User user = userRepository.findById(userId)
@@ -39,13 +39,48 @@ public class ScrapService {
         return scrapFolderRepository.save(folder).getFolderId();
     }
 
-    // scrap 생성
+    // 스크랩 폴더 조회
+    public List<ScrapFolderResponse> getScrapFolders(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("유저 x"));
+
+        // 유저가 만든 폴더들 다 찾아오기
+        return scrapFolderRepository.findByUser(user).stream()
+                .map(ScrapFolderResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    // 스크랩 폴더 수정
+    @Transactional
+    public void updateScrapFolder(Long userId, Long folderId, ScrapReqest.UpdateScrapFolder req) {
+        ScrapFolder folder = scrapFolderRepository.findById(folderId)
+                .orElseThrow(() -> new IllegalArgumentException("폴더가 없습니다."));
+
+        validateFolderOwnership(folder, userId);
+
+        folder.updateScrapFolder(req.getName(), req.getDescription(), req.getFolderImageUrl());
+    }
+
+    // 스크랩 폴더 삭제
+    @Transactional
+    public void deleteScrapFolder(Long userId, Long folderId) {
+        ScrapFolder folder = scrapFolderRepository.findById(folderId)
+                .orElseThrow(() -> new IllegalArgumentException("폴더가 없습니다."));
+
+        validateFolderOwnership(folder, userId);
+
+        // Soft Delete (@SQLDelete) 작동(자동으로)
+        scrapFolderRepository.delete(folder);
+    }
+
+    //--------------------------------------------------------------------------------------------
+    // 스크랩 생성
     @Transactional
     public Long createScrap(Long userId, Long folderId, ScrapReqest.CreateScrap req){
         ScrapFolder folder = scrapFolderRepository.findById(folderId)
                 .orElseThrow(() -> new IllegalArgumentException("폴더 x"));
 
-        // 내 폴더인지 확인 필요
+        validateFolderOwnership(folder, userId);
 
         Scrap scrap = Scrap.builder()
                 .scrapFolder(folder)
@@ -57,49 +92,17 @@ public class ScrapService {
         return scrapRepository.save(scrap).getScrapId();
     }
 
-    // scrap folder 조회
-    public List<ScrapFolderResponse> getScrapFolders(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("유저 x"));
-
-        // 유저가 만든 폴더들 다 찾아오기
-        return scrapFolderRepository.findByUser(user).stream()
-                .map(ScrapFolderResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    // scrap 조회
+    // 스크랩 조회
     @Transactional
-    public List<ScrapResponse> getScraps(Long folderId){
+    public List<ScrapResponse> getScraps(Long userId, Long folderId){
         ScrapFolder folder = scrapFolderRepository.findById(folderId)
                 .orElseThrow(() -> new IllegalArgumentException("폴더 없음"));
+
+        validateFolderOwnership(folder, userId);
 
         return scrapRepository.findByScrapFolder(folder).stream()
                 .map(ScrapResponse::from)
                 .collect(Collectors.toList());
-    }
-
-    // 폴더 수정
-    @Transactional
-    public void updateScrapFolder(Long userId, Long folderId, ScrapReqest.UpdateScrapFolder req) {
-        ScrapFolder folder = scrapFolderRepository.findById(folderId)
-                .orElseThrow(() -> new IllegalArgumentException("폴더가 없습니다."));
-
-        // user 통해서 폴더 주인 확인로직 추가 필요
-
-        folder.updateScrapFolder(req.getName(), req.getDescription(), req.getFolderImageUrl());
-    }
-
-    // 폴더 삭제
-    @Transactional
-    public void deleteScrapFolder(Long userId, Long folderId) {
-        ScrapFolder folder = scrapFolderRepository.findById(folderId)
-                .orElseThrow(() -> new IllegalArgumentException("폴더가 없습니다."));
-
-        // 권한 체크
-
-        // Soft Delete (@SQLDelete) 작동(자동으로)
-        scrapFolderRepository.delete(folder);
     }
 
     // 스크랩 수정
@@ -109,6 +112,7 @@ public class ScrapService {
                 .orElseThrow(() -> new IllegalArgumentException("스크랩이 없습니다."));
 
         // 권한 확인
+        validateScrapOwnership(scrap, userId);
 
         scrap.updateScrap(req.getTitle(), req.getContent(), req.getOriginalLink(), req.getImageUrl());
     }
@@ -119,8 +123,23 @@ public class ScrapService {
         Scrap scrap = scrapRepository.findById(scrapId)
                 .orElseThrow(() -> new IllegalArgumentException("스크랩이 없습니다."));
 
-        // User 통해서 폴더 주인 확인로직 추가 필요
+        validateScrapOwnership(scrap, userId);
 
         scrapRepository.delete(scrap);
+    }
+
+    //-------------------------------------------------------------------------------------------------
+    // 폴더의 주인이 현재 로그인한 유저인지 확인
+    private void validateFolderOwnership(ScrapFolder folder, Long userId) {
+        if (!folder.getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("해당 폴더에 대한 권한이 없습니다.");
+        }
+    }
+
+    // 스크랩 -> 폴더 -> 유저 : 로 확인
+    private void validateScrapOwnership(Scrap scrap, Long userId) {
+        if (!scrap.getScrapFolder().getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("해당 스크랩에 대한 권한이 없습니다.");
+        }
     }
 }
